# [XPU] Implemented 32bit optimizers in triton