--- a/bitsandbytes/functional.py
+++ b/bitsandbytes/functional.py
@@ -1199,9 +1199,9 @@ def optimizer_update_32bit(
         beta3,
         alpha,
         eps,
-        weight_decay,
         step,
         lr,
+        weight_decay,
         gnorm_scale,
         skip_zeros,
     )